change source structure, move navbar to app
