+++ src/App.js
@@ -1,12 +1,13 @@
 import React from 'react';
 import { Switch, Route } from 'react-router-dom'
 
-import Navbar from './MainNavbar.js';
-import Landing from'./Landing.js';
+import Navbar from './components/MainNavbar.js';
+import Landing from'./pages/Landing.js';
 
 function App () {
   return (
     <div className="App">
+      <Navbar />
       <Switch>
         <Route path="/" component={Landing} />
       </Switch>
